audio_analysis.py created. currenlty only contains get_sampling_rate(), used in asr for non-fixed sampling rate (now taken from file)
