+++ main.py
@@ -14,21 +14,6 @@ def load_reference_data(path):
     return text
 
 
-"""
-def silence_removal(audio_path):
-    audio, sample_rate = read_wave(audio_path)
-    vad = webrtcvad.Vad(int(1))
-    frames = frame_generator(30, audio, sample_rate)
-    frames = list(frames)
-    segments = vad_collector(sample_rate, 30, 300, vad, frames)
-
-    # Segmenting the Voice audio and save it in list as bytes
-    concataudio = [segment for segment in segments]
-    joinedaudio = b"".join(concataudio)
-    write_wave("Non-Silenced-Audio.wav", joinedaudio, sample_rate)
-"""
-
-
 def main():
     conf = load_conf_file('config.json')
